Refactor classes accordingly to review comments
